Refactor validate-config to happen at runtime
