+++ clients/privacy-center/next.config.js
@@ -1,7 +1,4 @@
 const path = require("path");
-const { validateConfig } = require("./scripts/validate-config.js");
-
-validateConfig();
 
 /** @type {import('next').NextConfig} */
 const withBundleAnalyzer = require("@next/bundle-analyzer")({
